Implement code review suggestions
