+++ src/ansys/acp/core/_tree_objects/__init__.py
@@ -20,8 +20,8 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-from ._mesh import MeshData
-from ._mesh_data import ScalarData, VectorData
+from ._elemental_or_nodal_data import ScalarData, VectorData
+from ._mesh_data import MeshData
 from .analysis_ply import AnalysisPly, AnalysisPlyElementalData, AnalysisPlyNodalData
 from .boolean_selection_rule import (
     BooleanSelectionRule,
