+++ src/ansys/acp/core/_tree_objects/analysis_ply.py
@@ -30,13 +30,7 @@ import numpy as np
 from ansys.api.acp.v0 import analysis_ply_pb2, analysis_ply_pb2_grpc
 
 from .._utils.property_protocols import ReadOnlyProperty
-from ._grpc_helpers.property_helper import (
-    grpc_data_property_read_only,
-    grpc_link_property_read_only,
-    mark_grpc_properties,
-)
-from ._mesh import mesh_property, shell_mesh_property, solid_mesh_property
-from ._mesh_data import (
+from ._elemental_or_nodal_data import (
     ElementalData,
     NodalData,
     ScalarData,
@@ -44,6 +38,12 @@ from ._mesh_data import (
     elemental_data_property,
     nodal_data_property,
 )
+from ._grpc_helpers.property_helper import (
+    grpc_data_property_read_only,
+    grpc_link_property_read_only,
+    mark_grpc_properties,
+)
+from ._mesh_data import full_mesh_property, shell_mesh_property, solid_mesh_property
 from .base import IdTreeObject, ReadOnlyTreeObject
 from .enums import status_type_from_pb
 from .object_registry import register
@@ -121,7 +121,7 @@ class AnalysisPly(ReadOnlyTreeObject, IdTreeObject):
         "properties.active_in_post_mode"
     )
 
-    mesh = mesh_property
+    mesh = full_mesh_property
     shell_mesh = shell_mesh_property
     solid_mesh = solid_mesh_property
     elemental_data = elemental_data_property(AnalysisPlyElementalData)
